moved GeneralResponse
